Phase 8.9c: drive UI model filter from registry
